refactor(session): split bootstrap.rs into submodules
Split 722-line bootstrap.rs into a bootstrap/ directory with the following modules:
- types.rs: BootstrapProtectionError, BootstrapRegion, and constants
- context.rs: BootstrapContext struct and impl
- protection.rs: BootstrapProtection struct and impl
- helpers.rs: make_bootstrap_marker utility
- tests.rs: all unit tests
- mod.rs: module declarations and re-exports

Source: issue #217

Type: refactor
